+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Extensions
   - `dear-imnodes`: `EditorContext::*` methods that rely on the global current ImNodes context are deprecated and will be removed in `0.11.0`. Use `Context::bind_editor(&EditorContext)` instead.
+  - `dear-imnodes`: `EditorContext::create/try_create` are deprecated and will be removed in `0.11.0`. Use `Context::{create_editor_context,try_create_editor_context}` instead.
 
 ## [0.10.2] - 2026-02-24